move request/response dataclasses to new file
